backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.27
huggingface-hub==0.24.6